none git pull git commit -m none
