+++ cishouseholds/pipeline/pipeline_stages.py
@@ -21,12 +21,12 @@ from cishouseholds.merge import union_dataframes_to_hive
 from cishouseholds.pipeline.category_map import category_maps
 from cishouseholds.pipeline.config import get_config
 from cishouseholds.pipeline.config import get_secondary_config
-from cishouseholds.pipeline.ETL_scripts import extract_validate_transform_input_data
 from cishouseholds.pipeline.generate_outputs import map_output_values_and_column_names
 from cishouseholds.pipeline.generate_outputs import write_csv_rename
 from cishouseholds.pipeline.load import check_table_exists
 from cishouseholds.pipeline.load import extract_df_list
 from cishouseholds.pipeline.load import extract_from_table
+from cishouseholds.pipeline.load import extract_validate_transform_input_data
 from cishouseholds.pipeline.load import get_full_table_name
 from cishouseholds.pipeline.load import update_table
 from cishouseholds.pipeline.load import update_table_and_log_source_files
